protected routes using JWT

+++ backend/controllers/taskController.js
@@ -1,10 +1,11 @@
 const asyncHandler = require('express-async-handler');
 const Task = require('../models/taskModel');
+const User = require('../models/userModel');
 const mongoose = require('mongoose');
 
 // GET
 const getTasks = asyncHandler(async (req, res) => {
-  const tasks = await Task.find();
+  const tasks = await Task.find({ user: req.user.id });
   res.status(200).json(tasks);
 });
 
@@ -14,12 +15,12 @@ const setTask = asyncHandler(async (req, res) => {
     res.status(400);
     throw new Error('Please enter a task');
   }
-  const task = await Task.create({ text: req.body.text });
+  const task = await Task.create({ text: req.body.text, user: req.user.id });
   res.status(200).json(task);
 });
 
 // PUT
-const updateTask = asyncHandler(async (req, res, next) => {
+const updateTask = asyncHandler(async (req, res) => {
   const taskId = req.params.id;
 
   if (!mongoose.isValidObjectId(taskId)) {
@@ -31,6 +32,18 @@ const updateTask = asyncHandler(async (req, res, next) => {
   if (!task) {
     res.status(400);
     throw new Error('Task not found');
+  }
+
+  // throwing an error if we do not find a user associated with the updated task or if the bearer token is wrong
+  const user = await User.findById(req.user.id);
+  if (!user) {
+    res.status(401);
+    throw new Error('No such user found');
+  }
+
+  if (task.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error('user is not authorized to update');
   }
 
   const updatedTask = await Task.findByIdAndUpdate(req.params.id, req.body, {
@@ -55,6 +68,16 @@ const deleteTask = asyncHandler(async (req, res) => {
     res.status(400);
     throw new Error('Task not found');
   }
+  const user = await User.findById(req.user.id);
+  if (!user) {
+    res.status(401);
+    throw new Error('No such user found');
+  }
+  if (task.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error('User is not authorized to update');
+  }
+
   await Task.findByIdAndDelete(req.params.id);
 
   res.status(200).json({ id: req.params.id });
